backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
sentence-transformers
numpy
pydantic
